Improvements after review. Related to Issue [WTI-773]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@7438 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: bf03f344571164795f315d8a063a3bc2bdc2b9c4
Former-commit-id: beb549c0f7dc4e2aeaee8e8152a840fa15364b86

+++ Ringtoets/GrassCoverErosionOutwards/test/Ringtoets.GrassCoverErosionOutwards.Plugin.Test/PropertyInfos/GrassCoverErosionOutwardsWaveConditionsOutputPropertiesTest.cs
@@ -68,7 +68,7 @@ namespace Ringtoets.GrassCoverErosionOutwards.Plugin.Test.PropertyInfos
 
             // Assert 
             CollectionAssert.AllItemsAreInstancesOfType(properties.Items, typeof(WaveConditionsOutputProperties));
-            Assert.AreEqual((int) items.Length, properties.Items.Length);
+            Assert.AreEqual(items.Length, properties.Items.Length);
 
             var firstOutput = items[0];
             var firstOutputProperties = properties.Items[0];
@@ -78,11 +78,11 @@ namespace Ringtoets.GrassCoverErosionOutwards.Plugin.Test.PropertyInfos
             Assert.AreEqual(firstOutput.WaveAngle, firstOutputProperties.WaveAngle);
             Assert.AreEqual(firstOutput.WaveDirection, firstOutputProperties.WaveDirection);
             Assert.AreEqual(ProbabilityFormattingHelper.Format(firstOutput.TargetProbability), firstOutputProperties.TargetProbability);
-            Assert.AreEqual((double) firstOutput.TargetReliability, (double) firstOutputProperties.TargetReliability,
-                            (double) firstOutputProperties.TargetReliability.GetAccuracy());
+            Assert.AreEqual(firstOutput.TargetReliability, firstOutputProperties.TargetReliability,
+                            firstOutputProperties.TargetReliability.GetAccuracy());
             Assert.AreEqual(ProbabilityFormattingHelper.Format(firstOutput.CalculatedProbability), firstOutputProperties.CalculatedProbability);
-            Assert.AreEqual((double) firstOutput.TargetReliability, (double) firstOutputProperties.TargetReliability,
-                            (double) firstOutputProperties.TargetReliability.GetAccuracy());
+            Assert.AreEqual(firstOutput.TargetReliability, firstOutputProperties.TargetReliability,
+                            firstOutputProperties.TargetReliability.GetAccuracy());
 
             string convergenceValue = new EnumDisplayWrapper<CalculationConvergence>(firstOutput.CalculationConvergence).DisplayName;
             Assert.AreEqual(convergenceValue, firstOutputProperties.Convergence);
